move server implementation to server package for pfs
