--- src/pfs/pfs.go
@@ -1,6 +1,0 @@
-package pfs
-
-// NewAPIServer returns a new ApiServer.
-func NewAPIServer() ApiServer {
-	return newAPIServer()
-}
